fix walletEid

+++ stratumsdk/walletAddresses.go
@@ -5,7 +5,7 @@ import (
 )
 
 type WalletAddressData struct {
-	WalletAddressEid   int     `json:"wallet_address_eid"`
+	WalletAddressEid   int64   `json:"wallet_address_eid"`
 	WalletEid          int     `json:"wallet_eid"`
 	WalletAddressLabel string  `json:"wallet_address_label"`
 	WalletAddress      string  `json:"wallet_address"`
